Move delimited literal to someplace useful
