+++ lib/matchers/lang/cfamily/javascript/regexpLiteral.ts
@@ -1,0 +1,6 @@
+import { MatchingLogic } from "../../../../Matchers";
+import { DelimitedLiteral } from "../DelimitedLiteral";
+
+export function regexLiteral(): MatchingLogic {
+    return new DelimitedLiteral("/");
+}
